requirements.txt
alembic==1.8.0
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
SQLAlchemy==1.4.37
watchdog==2.1.8
Werkzeug==2.1.2
zipp==3.8.0
